+++ EPOH/Assets/Scripts/Boss/4_Hoa/NeuronActive.cs
@@ -7,7 +7,7 @@ public class NeuronActive : MonoBehaviour
     [SerializeField] private Vector3[] neuronPosition;
     [SerializeField] private GameObject neuron;
     private HackingForN hacking;
-    private int hackingGoal;
+    private float hackingGoal;
 
     public static NeuronActive NM;
